requirements.txt
streamlit
torch
torchvision
opencv-python
Pillow
numpy
matplotlib
facenet-pytorch
grad-cam
